backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-import-export==3.3.8
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.13.1
html5lib==1.1
idna==3.7
lxml==5.2.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycparser==2.22
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.3.0
typing_extensions==4.11.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
virtualenv==20.25.1
webencodings==0.5.1
xhtml2pdf==0.2.15
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
paddlepaddle==2.6.0
paddleocr==2.6.0